+++ src/vue/pages/KycRecovery/KycRecoveryStateMessage.vue
@@ -124,7 +124,7 @@ export default {
 .kyc-recovery-state-message__content {
   padding: 2.4rem;
   font-size: 1.4rem;
-  color: $col-primary-txt;
+  color: $col-message-box-text;
   word-wrap: break-word;
   line-height: 1.25;
 }