+++ src/vue/pages/Verification.vue
@@ -9,11 +9,6 @@
         </p>
         <div class="account-type-selector">
           <router-link
-            v-if="
-              getModule().canRenderSubmodule(VerificationGeneralPageModule) ||
-                getModule()
-                  .canRenderSubmodule(VerificationGeneralAdvancedPageModule)
-            "
             tag="button"
             :to="vueRoutes.verificationGeneral"
             class="account-type-selector__item"
@@ -35,9 +30,6 @@
           </router-link>
 
           <router-link
-            v-if="
-              getModule().canRenderSubmodule(VerificationCorporatePageModule)
-            "
             tag="button"
             :to="vueRoutes.verificationCorporate"
             class="account-type-selector__item"
@@ -87,10 +79,6 @@ import { ErrorHandler } from '@/js/helpers/error-handler'
 
 import { REQUEST_STATES_STR } from '@/js/const/request-states.const'
 import config from '@/config'
-
-import { VerificationCorporatePageModule } from './verification-corporate-page-module'
-import { VerificationGeneralPageModule } from './verification-general-page-module'
-import { VerificationGeneralAdvancedPageModule } from './verification-general-advanced-page-module'
 
 // The guard doesn't allow the user to visit a verification page
 // if he/she has already sent the verification request, and the admin
@@ -147,9 +135,6 @@ export default {
     vueRoutes,
     REQUEST_STATES_STR,
     config,
-    VerificationCorporatePageModule,
-    VerificationGeneralPageModule,
-    VerificationGeneralAdvancedPageModule,
   }),
 
   computed: {
@@ -247,7 +232,7 @@ export default {
 }
 
 .router-link-exact-active {
-  border: 0.2rem solid $col-primary-lighten;
+  border: 0.2rem solid $col-btn;
 }
 
 .account-type-selector__item-title {
@@ -267,7 +252,7 @@ export default {
   height: 2.4rem;
   padding: 0.2rem;
   background-color: $col-block-bg;
-  border: 0.2rem solid $col-primary-lighten;
+  border: 0.2rem solid $col-btn;
   border-radius: 2rem;
   top: -1.2rem;
   right: -1.2rem;
@@ -281,7 +266,7 @@ export default {
 
 .account-type-selector__selected-icon-tag {
   font-size: 1.6rem;
-  color: $col-primary-lighten;
+  color: $col-btn;
 }
 
 .verification__form-label {